Protobuf is now compiled into the  executable and it runs.  Display configuration is changed to use the protobuf as it's storage internally.
Read/write functions are written, but currently unused.

I changed compile engine from Ninja to MSVC 2019 for my cmake instance because it seemed to work a bit better.  My hack of copying over the protobuf dll from vcpkg may not work for ninja on windows (and shouldn't be necessary on linux at all, but that's untested)