Added TextEntry which updates the previews (but not the presenters).
Fixed a bug with dll placement for the release build.

Fixed a bug with the WXDEBUG macro being present in release builds (turn off debugging in WX in release altogether, in essence)

Fixed a warning in DisplayConfig::primaryDisplay

Created a TextEntryTest, but the action driver doesn't appear to be working correctly yet, but submitting what I have and continuing to debug.