Started work on a text serialization option, but it's pretty broken at the moment (sadface)
Cleaned up some errors in serializing binary data and it now saves at the end of every display config.  We're just going with binary for the time being, as it's easier to manage.  This may make a future release that has to read either binary or text more complicated, but I'll deal with that later.