Some more Linux work.
I gave up on having protos in a subdirectory, the protobuf cmake methods
just really want them in the root of the project, so I've just ceded to
their demands.  Updated all references to protos and now the windows &
linux methods can reconverge.  A cleanup commit will follow this one.