Baby's first gmock
Broke out communication with the outside world into a separate class, mocked the class and moved testing over to the mock, except for one representative test to make sure that coverage is still about the same.

Added some stuff in Cmake to allow for an Integration Test flag that turns on more expensive tests, so I can work on mocking a lot more and getting tests faster and less dependent on environmental setup for regular testing.