Changing addWidget() to take a const ref instead of a pointer.
This has the benefit of making calls a few characters shorter in the normal case and also reduces the times we call .get() on unique_ptrs, which makes me happy in concept.

It also is a good idea to take const ref if you don't intend to change the object, so I'll be auditing other things moving forward to see if we can do that more often.