Yet another pivot!
Perl isn't a great long-term maintenance idea for anything I might want other people to contribute to, so let's look at making this a NodeJS backend.  I just moved perl to the side for now, since it's got a lot of functionality I want to refer to, and created a Node/Express project with no functionality yet.